requirements.txt
numpy
pytorch::pytorch
